# api/openapi.yaml
openapi: 3.0.3
info:
  title: MoonLightCloud API
  description: >-
    Device lifecycle + versioned sync ingestion with the first Sale-sync
    vertical slice: durable inbox, sale.finalized.v1 validated before ACK,
    async projection with durable processing state. Business APIs are
    versioned under /api/v1. Health and version stay outside the version
    boundary. ACK means durable PostgreSQL commit, never queueing alone.
  version: 2D
paths:
...
  /api/v1/dashboard/devices:
    get:
      summary: Operator device connectivity + latest sync command
      operationId: dashboardDevices
      security:
        - dashboardSession: []
      responses:
        '200':
          description: Devices with NEVER_SEEN/ONLINE/OFFLINE connectivity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardDeviceList'
        '401':
          description: Missing or invalid operator session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN (missing permission), STORE_FORBIDDEN / STORE_SCOPE_REQUIRED (Store membership), MFA_REQUIRED / MFA_SETUP_REQUIRED (session not FULL) or CSRF_REJECTED (mutations).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/dashboard/devices/{device_id}/sync-requests:
    post:
      summary: Queue sync_now.v1 (Idempotency-Key required, one active per device)
      operationId: dashboardSyncRequest
      security:
        - dashboardSession: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          description: Caller-generated idempotency key, unique per device. Missing or invalid keys are rejected with 400 DEVICE_COMMAND_INVALID_IDEMPOTENCY_KEY.
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^[A-Za-z0-9_:.~-]{1,128}$
      responses:
        '201':
          description: Command created (pending).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSyncRequestResponse'
        '200':
          description: Same-key idempotent replay returns the existing command with created=false.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSyncRequestResponse'
        '400':
          description: Invalid idempotency key or inactive device (DEVICE_COMMAND_INVALID_IDEMPOTENCY_KEY / DEVICE_NOT_ACTIVE).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid operator session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN (missing permission), STORE_FORBIDDEN / STORE_SCOPE_REQUIRED (Store membership), MFA_REQUIRED / MFA_SETUP_REQUIRED (session not FULL) or CSRF_REJECTED (mutations).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Contradictory reuse (DEVICE_COMMAND_CONFLICT) or another sync already active for this device (DEVICE_SYNC_ALREADY_ACTIVE). Returns only the error envelope; no command is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Malformed or unknown device id (DEVICE_COMMAND_NOT_FOUND).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    deviceAuth:
      type: http
      scheme: bearer
      bearerFormat: '<device-id>.<credential-id>.<secret-hex> (raw secret shown once at provisioning/rotation)'
      description: >-
        Authorization: Bearer <device_uuid>.<credential_uuid>.<64-hex secret>.
        Never in query parameters. HTTPS in production.
    reportAuth:
      type: http
      scheme: bearer
      bearerFormat: 'REPORTING_API_TOKEN (temporary report-read secret until dashboard user auth)'
      description: >-
        Authorization: Bearer <reporting-api-token>. Report scope only;
        device sync credentials are never valid here. Fail-closed: required
        16+ chars in staging/production; open reports need ALL of
        ENVIRONMENT=development, ALLOW_UNAUTHENTICATED_REPORTING=true, and no
        token. Anything else (including missing ENVIRONMENT) fails startup.
    dashboardSession:
      type: apiKey
      in: cookie
      name: __Host-mlc_session
      description: >-
        Opaque server-side human session (ADR-0053): 256-bit random token,
        stored only as a digest. Cookie __Host-mlc_session (HttpOnly, Secure,
        SameSite=Strict, Path=/) outside development; mlc_session over plain
        HTTP in development. Every dashboard route also requires the FULL
        (MFA-complete) stage, a named permission and server-side Store
        membership; mutations additionally require a same-origin Origin/Referer
        and the X-CSRF-Token header. Device credentials and
        REPORTING_API_TOKEN are never valid here.
...
  schemas:
...
    DeviceControlCommand:
      type: object
      description: Durable sync_now.v1 command. Only command_type sync_now and command_version 1 exist.
      required: [id, type, version, lease_generation, requested_at, status]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [sync_now]
        version:
          type: integer
          enum: [1]
        lease_generation:
          type: integer
          minimum: 0
          description: Increments on every poll lease; diagnostics only.
        requested_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, leased, accepted, running, completed, failed]
        accepted_at:
          type: string
          format: date-time
          nullable: true
        running_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        result_code:
          type: string
          nullable: true
          enum: [SYNC_COMPLETED, SYNC_FAILED_NETWORK, SYNC_FAILED_AUTH, SYNC_FAILED_CONFLICT, SYNC_FAILED_INTERNAL]
...
    DashboardDeviceList:
      type: object
      required: [devices]
      properties:
        devices:
          type: array
          items:
            $ref: '#/components/schemas/DashboardDevice'
    DashboardDevice:
      type: object
      description: Operator device row. Connectivity derives from server-observed last_seen_at; active_command is null unless a non-terminal command exists; recent_commands holds up to 5 newest commands of any state.
      required: [device_id, lifecycle, connectivity]
      properties:
        device_id:
          type: string
          format: uuid
        name:
          type: string
        lifecycle:
          type: string
          description: Canonical device lifecycle (active or revoked).
        connectivity:
          type: string
          enum: [NEVER_SEEN, ONLINE, OFFLINE]
          description: ONLINE means the control plane heard from the Retail app within the freshness window; never proof of data freshness.
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        active_command:
          $ref: '#/components/schemas/DeviceControlCommand'
          nullable: true
        recent_commands:
          type: array
          items:
            $ref: '#/components/schemas/DeviceControlCommand'
        store_id:
          type: string
          format: uuid
          description: Bound Store, if any. Absent means unbound/legacy — never fabricated.
        store_name:
          type: string
          description: Bound Store display name, if any.
...
    DashboardSyncRequestResponse:
      type: object
      required: [command, created]
      properties:
        command:
          $ref: '#/components/schemas/DeviceControlCommand'
        created:
          type: boolean
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: UNAUTHORIZED
            message:
              type: string
              example: invalid device credential
...